cursor_plan/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=24.2,<25.0
python-dateutil>=2.9,<3.0
numpy>=1.26,<3.0
pandas>=2.2,<3.0
scikit-learn>=1.5,<2.0
shapely>=2.0,<3.0
geopy>=2.4,<3.0
orjson>=3.10,<4.0
whitenoise>=6.7,<7.0
django-cas-ng>=5.0,<6.0
